Don't hang onto expired resources without validation headers in memory cache https://bugs.webkit.org/show_bug.cgi?id=226083 rdar://78035612
Reviewed by Chris Dumez.

They consume memory while only really being useful for history navigation.
Disk cache can handle the relatively rare cases where that is useful and not
covered by the page cache.

* loader/ResourceTimingInformation.cpp:
(WebCore::ResourceTimingInformation::addResourceTiming):
(WebCore::ResourceTimingInformation::removeResourceTiming):

Add way to remove entries from the map.

* loader/ResourceTimingInformation.h:
* loader/cache/CachedResource.cpp:
(WebCore::CachedResource::removeClient):

Removal from cache is now handled in deleteIfPossible() so any path that gets the resource to
canDelete() state will do cache removal check.
Since deleteIfPossible is invoked by removeClient the existing behavior is covered.

(WebCore::CachedResource::deleteIfPossible):

Remove expired resources without validation headers from the cache if the resource is otherwise deletable.
Also eagerly remove no-store http resources (previously only https was removed).

(WebCore::CachedResource::deleteThis):

Factor into a function.

* loader/cache/CachedResource.h:
* loader/cache/CachedResourceLoader.cpp:
(WebCore::CachedResourceLoader::requestResource):

Ensure existing resource is removed from cache adding a new one. This was causing confusion with inCache state.

(WebCore::CachedResourceLoader::garbageCollectDocumentResources):

Don't remove entries with active loader. They are needed for resource timing.
Don't leave removed entries to ResourceTimingInformation. They are not guaranteed to stay alive after
being removed from m_documentResources.

* loader/cache/MemoryCache.cpp:
(WebCore::MemoryCache::remove):

Call CachedResource::deleteThis() directly.


Canonical link: https://commits.webkit.org/238169@main
git-svn-id: https://svn.webkit.org/repository/webkit/trunk@278119 268f45cc-cd09-0410-ab3c-d52691b4dbfc

+++ Source/WebCore/loader/cache/CachedResource.cpp
@@ -552,11 +552,6 @@ void CachedResource::removeClient(CachedResourceClient& client)
         didRemoveClient(client);
     }
 
-    if (deleteIfPossible()) {
-        // `this` object is dead here.
-        return;
-    }
-
     if (hasClients())
         return;
 
@@ -565,6 +560,12 @@ void CachedResource::removeClient(CachedResourceClient& client)
         memoryCache.removeFromLiveResourcesSize(*this);
         memoryCache.removeFromLiveDecodedResourcesList(*this);
     }
+
+    if (deleteIfPossible()) {
+        // `this` object is dead here.
+        return;
+    }
+
     if (!m_switchingClientsToRevalidatedResource)
         allClientsRemoved();
     destroyDecodedDataIfNeeded();
@@ -572,13 +573,6 @@ void CachedResource::removeClient(CachedResourceClient& client)
     if (!allowsCaching())
         return;
 
-    if (response().cacheControlContainsNoStore() && url().protocolIs("https")) {
-        // RFC2616 14.9.2:
-        // "no-store: ... MUST make a best-effort attempt to remove the information from volatile storage as promptly as possible"
-        // "... History buffers MAY store such responses as part of their normal operation."
-        // We allow non-secure content to be reused in history, but we do not allow secure content to be reused.
-        memoryCache.remove(*this);
-    }
     memoryCache.pruneSoon();
 }
 
@@ -604,19 +598,47 @@ void CachedResource::decodedDataDeletionTimerFired()
 
 bool CachedResource::deleteIfPossible()
 {
-    if (canDelete()) {
-        LOG(ResourceLoading, "CachedResource %p deleteIfPossible - can delete, in cache %d", this, inCache());
-        if (!inCache()) {
-            InspectorInstrumentation::willDestroyCachedResource(*this);
-            delete this;
-            return true;
-        }
-        if (m_data)
-            m_data->hintMemoryNotNeededSoon();
+    if (!canDelete()) {
+        LOG(ResourceLoading, "CachedResource %p deleteIfPossible - can't delete (hasClients %d loader %p preloadCount %u handleCount %u resourceToRevalidate %p proxyResource %p)", this, hasClients(), m_loader.get(), m_preloadCount, m_handleCount, m_resourceToRevalidate, m_proxyResource);
+        return false;
     }
 
-    LOG(ResourceLoading, "CachedResource %p deleteIfPossible - can't delete (hasClients %d loader %p preloadCount %u handleCount %u resourceToRevalidate %p proxyResource %p)", this, hasClients(), m_loader.get(), m_preloadCount, m_handleCount, m_resourceToRevalidate, m_proxyResource);
+    LOG(ResourceLoading, "CachedResource %p deleteIfPossible - can delete, in cache %d", this, inCache());
+
+    if (!inCache()) {
+        deleteThis();
+        return true;
+    }
+
+    auto shouldRemoveFromCache = [&] {
+        // We may still keeps some of these cases in disk cache for history navigation.
+        if (response().cacheControlContainsNoStore())
+            return true;
+        if (isExpired() && !canUseCacheValidator())
+            return true;
+        return false;
+    }();
+
+    if (shouldRemoveFromCache) {
+        // Deletes this.
+        MemoryCache::singleton().remove(*this);
+        return true;
+    }
+
+    if (m_data)
+        m_data->hintMemoryNotNeededSoon();
+
     return false;
+}
+
+void CachedResource::deleteThis()
+{
+    RELEASE_ASSERT(canDelete());
+    RELEASE_ASSERT(!inCache());
+
+    InspectorInstrumentation::willDestroyCachedResource(*this);
+
+    delete this;
 }
 
 void CachedResource::setDecodedSize(unsigned size)
